backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
PyGithub
GitPython
radon
networkx
openai
python-dotenv
httpx
requests
sqlalchemy
passlib[bcrypt]
pyjwt
python-jose[cryptography]
chromadb
